InformSecondaryAdmin/requirements.txt
google-cloud-logging==2.4.0
SQLAlchemy==1.4.29
mailjet_rest==1.3.4
pg8000==1.23.0
